+++ src/gemseo_mlearning/active_learning/acquisition_criteria/maximum/_expected_impovement.py
@@ -83,8 +83,9 @@ class ExpectedImprovement:
             samples = self._compute_samples(
                 input_data=q_input_values, n_samples=self._mc_size
             )[..., 0]
-            improvement = (samples - self._qoi) * self._SIGN
-            return mean(np_max(maximum(improvement, 0), axis=0)) / self._scaling_factor
-        # distribution is not positive definite.
         except TypeError:
+            # The covariance matrix is not positive definite.
             return 0.0
+
+        improvement = (samples - self._qoi) * self._SIGN
+        return mean(np_max(maximum(improvement, 0), axis=0)) / self._scaling_factor