+++ docs/examples/active_learning/optimization/optim_batch.py
@@ -18,8 +18,7 @@
 from __future__ import annotations
 
 import matplotlib.pyplot as plt
-from gemseo import configure
-from gemseo import configure_logger
+from gemseo import configuration
 from gemseo import sample_disciplines
 from gemseo.mlearning.regression.algos.ot_gpr import OTGaussianProcessRegressor
 from numpy import argmin
@@ -32,10 +31,8 @@ from gemseo_mlearning.problems.rosenbrock.rosenbrock_discipline import (
 )
 from gemseo_mlearning.problems.rosenbrock.rosenbrock_space import RosenbrockSpace
 
-# Update the configuration of |g| to speed up the script (use configure() with care)
-configure(False, False, True, False, False, False, False)
-
-configure_logger()
+# Update the configuration of |g| to speed up the script.
+configuration.fast = True
 
 # %%
 # The use of active learning methods