+++ src/gemseo_mlearning/problems/rosenbrock/rosenbrock_discipline.py
@@ -62,6 +62,7 @@ class RosenbrockDiscipline(Discipline):
                 for input_name, derivative in zip(
                     self.io.input_grammar.names,
                     compute_gradient(inputs_array),
+                    strict=False,
                 )
             }
         }